requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.4.0
boto3==1.14.20
botocore==1.17.20
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.0.3
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-crispy-forms==1.9.1
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-storages==1.9.1
django-timezone-field==4.2.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
kombu==5.1.0
numpy==1.21.1
pandas==1.3.0
Pillow==7.1.2
prompt-toolkit==3.0.19
psycopg2==2.8.5
pyclean==2.0.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.1.0
